requirements_dev.txt
pip==19.2.3
bump2version==0.5.11
wheel==0.33.6
watchdog==0.9.0
flake8==3.7.8
tox==3.14.0
coverage==4.5.4
Sphinx==1.8.5
twine==1.14.0
Click==7.1.2
pytest==6.2.4
black==21.7b0
cycler==0.10.0
kiwisolver==1.3.1
klayout==0.27.0
matplotlib==3.4.2
numpy==1.19.3
Pillow==8.2.0
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.0
python-dateutil==2.8.1
scipy==1.6.3
PyYAML==5.3.1
six==1.15.0
